% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{.app_server}
\alias{.app_server}
\title{The Main App Server Function}
\usage{
.app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Shiny internal variables.}
}
\value{
Shiny reactive updates.
}
\description{
The Main App Server Function
}
\keyword{internal}
