% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{.get_questions}
\alias{.get_questions}
\title{Get R4DS Question Threads}
\usage{
.get_questions(question_channels)
}
\value{
A tidy tibble of question information.
}
\description{
Get R4DS Question Threads
}
\keyword{internal}
