% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{.render_questions}
\alias{.render_questions}
\title{Render the Questions DataTable}
\usage{
.render_questions(df)
}
\arguments{
\item{df}{A dataframe of questions.}
}
\value{
A \code{\link[DT]{renderDataTable}}.
}
\description{
Render the Questions DataTable
}
\keyword{internal}
