% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{.tidy_convos}
\alias{.tidy_convos}
\title{Rectangle Conversation Data}
\usage{
.tidy_convos(convos)
}
\arguments{
\item{convos}{A list returned by slackthreads::conversations.}

\item{question_channels}{A named character vector, where the name is the
user-friendly name of the channel, and the internal character is the
channel ID.}
}
\value{
A tibble of question data.
}
\description{
Rectangle Conversation Data
}
\keyword{internal}
