% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{.ui_header}
\alias{.ui_header}
\title{The Dashboard Header}
\usage{
.ui_header(dashboard_title)
}
\value{
A \code{\link[shinydashboard]{dashboardHeader}}.
}
\description{
The Dashboard Header
}
\keyword{internal}
