% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mentordash-package.R
\docType{package}
\name{mentordash-package}
\alias{mentordash}
\alias{mentordash-package}
\title{mentordash: Dashboard for the Data Science Learning Community}
\description{
This is a golem shiny app for the Data Science Learning Community.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r4ds/mentordash}
  \item Report bugs at \url{https://github.com/r4ds/mentordash/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com}

Other contributors:
\itemize{
  \item Tan Ho \email{tan@tanho.ca} [contributor]
}

}
\keyword{internal}
