% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(local = interactive(), ...)
}
\arguments{
\item{local}{A logical indicating whether you are launching the app locally,
rather than on shinyapps.io. Useful for debugging toward a deployed version
of this app.}
}
\description{
This function launches a Shiny App to help mentors find unanswered questions.
}
