% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{.is_answerable}
\alias{.is_answerable}
\title{Check Whether a Question is Answerable}
\usage{
.is_answerable(speech_balloons, users, reply_userses, channel_ids, tses)
}
\arguments{
\item{speech_balloons}{The logical vector indicating whether a question is
tagged with the "needs more information" emoji.}

\item{users}{The character vector of users who posted the question.}

\item{reply_userses}{The list of character vectors of users who have replied
to this thread.}

\item{tses}{The character vector of timestamps for threads attached to the
question.}

\item{channels}{The character vector of channels in which the questions were
posted.}
}
\value{
A logical vector.
}
\description{
Check Whether a Question is Answerable
}
\keyword{internal}
