% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{.is_answerable}
\alias{.is_answerable}
\title{Check Whether a Question is Answerable}
\usage{
.is_answerable(
  speech_balloons,
  users,
  reply_counts,
  reply_userses,
  channel_ids,
  tses,
  latest_activities,
  slack_api_key
)
}
\arguments{
\item{speech_balloons}{The logical vector indicating whether a question is
tagged with the "needs more information" emoji.}

\item{users}{The character vector of users who posted the question.}

\item{reply_counts}{The integer vector of numbers of replies to each message.}

\item{reply_userses}{The list of character vectors of users who have replied
to this thread.}

\item{channel_ids}{The character vector of channels in which the questions
were posted.}

\item{tses}{The character vector of timestamps for threads attached to the
question.}

\item{latest_activities}{The character vector of timestamps of the most
recent activity on each post.}

\item{slack_api_key}{This user's Slack token.}
}
\value{
A logical vector.
}
\description{
Check Whether a Question is Answerable
}
\keyword{internal}
