% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{.slack_shiny_ui}
\alias{.slack_shiny_ui}
\title{Require SLACK login to a Shiny app}
\usage{
.slack_shiny_ui(ui, team)
}
\arguments{
\item{ui}{A function defining the UI of the Shiny app, or a
\code{\link[shiny]{tagList}}.}

\item{team}{A Slack team ID for which the user should be authenticated.}
}
\value{
A function defining the UI of a Shiny app (either with login or
  without).
}
\description{
If the user hasn't logged in yet, return the login page. Otherwise return the
actual ui.
}
\keyword{internal}
