% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{.tidy_convos}
\alias{.tidy_convos}
\title{Rectangle Conversation Data}
\usage{
.tidy_convos(convos, slack_api_key)
}
\arguments{
\item{convos}{A list returned by slackthreads::conversations.}

\item{slack_api_key}{This user's Slack token.}
}
\value{
A tibble of question data.
}
\description{
Rectangle Conversation Data
}
\keyword{internal}
