% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{arguments to pass to golem_opts. See
\code{\link[golem]{get_golem_options}} for more details.}
}
\description{
This function launches a Shiny App to help mentors find unanswered questions.
}
