% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(...)
}
\description{
This function launches a Shiny App to help R4DS Online Learning Community
mentors find unanswered questions.
}
